+++ CHANGELOG.md
@@ -23,3 +23,4 @@ All notable changes to this project will be documented in this file.
 - Jobs API endpoints
 - Custom Sign In page
 - Sign In button animation
+- Toggle password visibility functionallity